Removed parallel implementation.
My first attempt at a naive implementation was no faster than the serial
implementation. In fact, it took generally twice as long. Research
suggests that this is a difficult algorithm to efficiently parallelize.
at 400ms for over 2 million data points, there doesn't seem to be a huge
case for parallelizing parts of the algorithm.